+++ frontend/src/pages/Dashboard/Footer.tsx
@@ -1,3 +1,7 @@
 export default function Footer() {
-  return <div>Footer</div>
+  return (
+    <div className="w-full h-40 bg-slate-300 flex flex-row justify-center items-center">
+      <div>Footer</div>
+    </div>
+  )
 }
