+++ frontend/src/pages/Dashboard/index.tsx
@@ -6,12 +6,14 @@ import PopularSeries from "./PopularSeries"
 
 export default function Dashboard() {
   return (
-    <div className="w-full md:px-10 space-y-10">
-      <PopularPublishers />
-      <PopularComics />
-      <PopularSeries />
-      <BrowseByEra />
+    <>
+      <div className="w-full md:px-10 space-y-10 mb-10">
+        <PopularPublishers />
+        <PopularComics />
+        <PopularSeries />
+        <BrowseByEra />
+      </div>
       <Footer />
-    </div>
+    </>
   )
 }
